+++ example/hello-mvc/webapp/main.dart
@@ -1,6 +1,7 @@
 //Sample of Stream: Hello Templates
 library hello_mvc;
 
+import "dart:async";
 import "dart:io";
 import "package:stream/stream.dart";
 
@@ -16,18 +17,23 @@ class FileInfo {
 }
 
 ///Controller: prepare the model and then invoke the view, listView
-void helloMVC(HttpConnect connect) {
+Future helloMVC(HttpConnect connect) {
   //1. prepare the model
+  final completer = new Completer();
   final curdir = new Directory.current();
   List<FileInfo> list = [];
 
   curdir.list().listen((fse) {
     list.add(new FileInfo(fse.path, fse is Directory));
   })
-  ..onError(connect.error)
+  ..onError((err) => completer.completeError(err))
   ..onDone(() {
-    listView(connect, path: curdir.path, infos: list); //forward to the view
+    listView(connect, path: curdir.path, infos: list).then((_) { //forward to the view
+      completer.complete();
+    }).catchError((err) => completer.completeError(err));
+    //TODO: if Stream.done is supported (issue 9725), we don't need completer
   });
+  return completer.future;
 }
 
 void main() {