+++ example/hello-rsp/webapp/main.dart
@@ -1,6 +1,7 @@
 //Sample of Stream: Hello RSP
 library hello_rsp;
 
+import "dart:async";
 import "dart:io";
 import "package:stream/stream.dart";