+++ lib/src/rspc/tagutil.dart
@@ -37,7 +37,7 @@ String toEL(String data, {direct: true}) {
       if (direct && i == 0 && j + 1 == len) //single EL
         return val;
       if (!val.isEmpty)
-        sb..write("\${nnstr(")..write(val)..write(")}");
+        sb..write("\${\$nns(")..write(val)..write(")}");
 
       i = j;
       continue;