+++ test/features/webapp/searchResult.rsp.dart
@@ -3,7 +3,7 @@
 part of features;
 
 /** Template, searchResult, for rendering the view. */
-void searchResult(HttpConnect connect, {criteria}) { //2
+Future searchResult(HttpConnect connect, {criteria}) { //#2
   var _cs_ = new List<HttpConnect>(), request = connect.request, response = connect.response;
 
   if (!connect.isIncluded)
@@ -21,28 +21,28 @@ void searchResult(HttpConnect connect, {criteria}) { //2
     <ul>
       <li>text: """); //#2
 
-  response.write(nnstr(criteria.text)); //#11
+  response.write($nns(criteria.text)); //#11
 
 
   response.write("""
 </li>
       <li>since: """); //#11
 
-  response.write(nnstr(criteria.since)); //#12
+  response.write($nns(criteria.since)); //#12
 
 
   response.write("""
 </li>
       <li>within: """); //#12
 
-  response.write(nnstr(criteria.within)); //#13
+  response.write($nns(criteria.within)); //#13
 
 
   response.write("""
 </li>
       <li>hasAttachment: """); //#13
 
-  response.write(nnstr(criteria.hasAttachment)); //#14
+  response.write($nns(criteria.hasAttachment)); //#14
 
 
   response.write("""
@@ -52,5 +52,5 @@ void searchResult(HttpConnect connect, {criteria}) { //2
 </html>
 """); //#14
 
-  connect.close();
+  return $nnf();
 }